Move icon variants to button
